requirements.txt
Flask==2.2.5
gunicorn==20.1.0
sentence-transformers==2.2.2
faiss-cpu==1.13.0
numpy==1.24.0
pandas==2.0.3
pymongo==4.4.0
python-dotenv==1.0.0
